+++ src/OpenSimCreator/UI/Shared/ModelSelectionGizmo.cpp
@@ -10,6 +10,7 @@
 #include <OpenSim/Simulation/Model/Model.h>
 #include <OpenSim/Simulation/Model/PathPoint.h>
 #include <OpenSim/Simulation/Model/PhysicalOffsetFrame.h>
+#include <OpenSim/Simulation/SimbodyEngine/Joint.h>
 #include <OpenSim/Simulation/Wrap/WrapObject.h>
 #include <oscar/Maths/Angle.h>
 #include <oscar/Maths/Eulers.h>
@@ -26,10 +27,10 @@
 #include <oscar/Maths/VecFunctions.h>
 #include <oscar/Platform/Log.h>
 #include <oscar/Shims/Cpp23/utility.h>
-#include <oscar/UI/ImGuizmoHelpers.h>
 #include <oscar/UI/oscimgui.h>
 #include <oscar/Utils/Assertions.h>
 #include <oscar/Utils/ScopeGuard.h>
+#include <oscar/Utils/Typelist.h>
 
 #include <concepts>
 #include <cstddef>
@@ -61,7 +62,7 @@ namespace
         return (cpp23::to_underlying(lhs) & cpp23::to_underlying(rhs)) != 0u;
     }
 
-    // type-erased virtual base class that each concrete manipulator inherits from
+    // type-erased interface to an object that manipulates something in a model
     class ISelectionManipulator {
     protected:
         ISelectionManipulator() = default;
@@ -77,19 +78,19 @@ namespace
             return implGetSupportedManipulationOps();
         }
 
-        Mat4 getCurrentModelMatrix() const
+        Mat4 getCurrentTransformInGround() const
         {
-            return implGetCurrentModelMatrix();
+            return implGetCurrentTransformInGround();
         }
 
-        void onApplyTranslation(const Vec3& deltaTranslationInGround)
+        void onApplyTranslation(const Vec3& translationInGround)
         {
-            implOnApplyTranslation(deltaTranslationInGround);
+            implOnApplyTranslation(translationInGround);
         }
 
-        void onApplyRotation(const Eulers& deltaEulerRadiansInGround)
+        void onApplyRotation(const Eulers& eulersInLocalSpace)
         {
-            implOnApplyRotation(deltaEulerRadiansInGround);
+            implOnApplyRotation(eulersInLocalSpace);
         }
 
         void onSave()
@@ -98,33 +99,38 @@ namespace
         }
     private:
         virtual SupportedManipulationOpFlags implGetSupportedManipulationOps() const = 0;
-        virtual Mat4 implGetCurrentModelMatrix() const = 0;
+        virtual Mat4 implGetCurrentTransformInGround() const = 0;
         virtual void implOnApplyTranslation(const Vec3&) {}  // default to noop
         virtual void implOnApplyRotation(const Eulers&) {}  // default to noop
         virtual void implOnSave() = 0;
     };
 
-    // concrete implementation of a selection manipulator for `TComponent`
-    //
-    // effectively, only stores the model+path to the thing being manipulated, and performs
-    // runtime checks to ensure the component still exists in the model
+    // abstract implementation of an `ISelectionManipulator` for the
+    // given `OpenSim::Component` (`TComponent`)
     template<std::derived_from<OpenSim::Component> TComponent>
     class SelectionManipulator : public ISelectionManipulator {
+    public:
+        using AssociatedComponent = TComponent;
+
+        static bool matches(const OpenSim::Component& component)
+        {
+            return dynamic_cast<const AssociatedComponent*>(&component);
+        }
     protected:
         SelectionManipulator(
             std::shared_ptr<UndoableModelStatePair> model_,
-            const TComponent& component) :
+            const AssociatedComponent& component) :
 
             m_Model{std::move(model_)},
             m_ComponentAbsPath{component.getAbsolutePath()}
         {
             OSC_ASSERT(m_Model != nullptr);
-            OSC_ASSERT(FindComponent<TComponent>(m_Model->getModel(), m_ComponentAbsPath));
+            OSC_ASSERT(FindComponent<AssociatedComponent>(m_Model->getModel(), m_ComponentAbsPath));
         }
 
-        const TComponent* findSelection() const
+        const AssociatedComponent* findSelection() const
         {
-            return FindComponent<TComponent>(m_Model->getModel(), m_ComponentAbsPath);
+            return FindComponent<AssociatedComponent>(m_Model->getModel(), m_ComponentAbsPath);
         }
 
         const OpenSim::Model& getModel() const
@@ -143,54 +149,56 @@ namespace
         }
 
     private:
-        // perform runtime lookup for `TComponent` and forward into concrete implementation
-        Mat4 implGetCurrentModelMatrix() const final
+        // perform runtime lookup for `AssociatedComponent` and forward into concrete implementation
+        Mat4 implGetCurrentTransformInGround() const final
         {
-            const TComponent* maybeSelected = findSelection();
-            if (!maybeSelected)
-            {
+            const AssociatedComponent* maybeSelected = findSelection();
+            if (not maybeSelected) {
                 return identity<Mat4>();  // selection of that type does not exist in the model
             }
-            return implGetCurrentModelMatrix(*maybeSelected);
+            return implGetCurrentTransformInGround(*maybeSelected);
         }
 
-        // perform runtime lookup for `TComponent` and forward into concrete implementation
-        void implOnApplyTranslation(const Vec3& deltaTranslationInGround) final
+        // perform runtime lookup for `AssociatedComponent` and forward into concrete implementation
+        void implOnApplyTranslation(const Vec3& translationInGround) final
         {
-            const TComponent* maybeSelected = findSelection();
-            if (!maybeSelected)
-            {
+            const AssociatedComponent* maybeSelected = findSelection();
+            if (not maybeSelected) {
                 return;  // selection of that type does not exist in the model
             }
-            implOnApplyTranslation(*maybeSelected, deltaTranslationInGround);
+            implOnApplyTranslation(*maybeSelected, translationInGround);
         }
 
-        // perform runtime lookup for `TComponent` and forward into concrete implementation
-        void implOnApplyRotation(const Eulers& deltaEulerRadiansInGround) final
+        // perform runtime lookup for `AssociatedComponent` and forward into concrete implementation
+        void implOnApplyRotation(const Eulers& eulersInLocalSpace) final
         {
-            const TComponent* maybeSelected = findSelection();
-            if (!maybeSelected)
-            {
+            const AssociatedComponent* maybeSelected = findSelection();
+            if (not maybeSelected) {
                 return;  // selection of that type does not exist in the model
             }
-            implOnApplyRotation(*maybeSelected, deltaEulerRadiansInGround);
+            implOnApplyRotation(*maybeSelected, eulersInLocalSpace);
         }
 
         void implOnSave() final
         {
-            const TComponent* maybeSelected = findSelection();
-            if (!maybeSelected)
-            {
+            const AssociatedComponent* maybeSelected = findSelection();
+            if (not maybeSelected) {
                 return;  // selection of that type does not exist in the model
             }
             implOnSave(*maybeSelected);
         }
 
+        virtual void implOnSave(const AssociatedComponent& component)
+        {
+            std::stringstream ss;
+            ss << "transformed " << component.getName();
+            getUndoableModel().commit(std::move(ss).str());
+        }
+
         // inheritors must implement concrete manipulation methods
-        virtual Mat4 implGetCurrentModelMatrix(const TComponent&) const = 0;
-        virtual void implOnApplyTranslation(const TComponent&, const Vec3& deltaTranslationInGround) = 0;
-        virtual void implOnApplyRotation(const TComponent&, const Eulers&) {}  // default to noop
-        virtual void implOnSave(const TComponent&) = 0;
+        virtual Mat4 implGetCurrentTransformInGround(const AssociatedComponent&) const = 0;
+        virtual void implOnApplyTranslation(const AssociatedComponent&, const Vec3& translationInGround) = 0;
+        virtual void implOnApplyRotation(const AssociatedComponent&, [[maybe_unused]] const Eulers& eulersInLocalSpace) {}
 
         std::shared_ptr<UndoableModelStatePair> m_Model;
         OpenSim::ComponentPath m_ComponentAbsPath;
@@ -200,7 +208,7 @@ namespace
 // concrete manipulator implementations
 namespace
 {
-    // manipulator for `OpenSim::Station`
+    // an `ISelectionManipulator` that manipulates an `OpenSim::Station`
     class StationManipulator final : public SelectionManipulator<OpenSim::Station> {
     public:
         StationManipulator(
@@ -210,29 +218,28 @@ namespace
         {}
 
     private:
-
         SupportedManipulationOpFlags implGetSupportedManipulationOps() const final
         {
             return SupportedManipulationOpFlags::Translation;
         }
 
-        Mat4 implGetCurrentModelMatrix(
+        Mat4 implGetCurrentTransformInGround(
             const OpenSim::Station& station) const final
         {
             const SimTK::State& state = getState();
-            Mat4 currentXformInGround = mat4_cast(station.getParentFrame().getRotationInGround(state));
-            currentXformInGround[3] = Vec4{ToVec3(station.getLocationInGround(state)), 1.0f};
+            Mat4 transformInGround = mat4_cast(station.getParentFrame().getRotationInGround(state));
+            transformInGround[3] = Vec4{ToVec3(station.getLocationInGround(state)), 1.0f};
 
-            return currentXformInGround;
+            return transformInGround;
         }
 
         void implOnApplyTranslation(
             const OpenSim::Station& station,
-            const Vec3& deltaTranslationInGround) final
+            const Vec3& translationInGround) final
         {
             const SimTK::Rotation parentToGroundRotation = station.getParentFrame().getRotationInGround(getState());
             const SimTK::InverseRotation& groundToParentRotation = parentToGroundRotation.invert();
-            const Vec3 translationInParent = ToVec3(groundToParentRotation * ToSimTKVec3(deltaTranslationInGround));
+            const Vec3 translationInParent = ToVec3(groundToParentRotation * ToSimTKVec3(translationInGround));
 
             ActionTranslateStation(getUndoableModel(), station, translationInParent);
         }
@@ -243,7 +250,7 @@ namespace
         }
     };
 
-    // manipulator for `OpenSim::PathPoint`
+    // an `ISelectionManpulator` that manipulates an `OpenSim::PathPoint`
     class PathPointManipulator : public SelectionManipulator<OpenSim::PathPoint> {
     public:
         PathPointManipulator(
@@ -258,23 +265,23 @@ namespace
             return SupportedManipulationOpFlags::Translation;
         }
 
-        Mat4 implGetCurrentModelMatrix(
+        Mat4 implGetCurrentTransformInGround(
             const OpenSim::PathPoint& pathPoint) const final
         {
             const SimTK::State& state = getState();
-            Mat4 currentXformInGround = mat4_cast(pathPoint.getParentFrame().getRotationInGround(state));
-            currentXformInGround[3] = Vec4{ToVec3(pathPoint.getLocationInGround(state)), 1.0f};
+            Mat4 transformInGround = mat4_cast(pathPoint.getParentFrame().getRotationInGround(state));
+            transformInGround[3] = Vec4{ToVec3(pathPoint.getLocationInGround(state)), 1.0f};
 
-            return currentXformInGround;
+            return transformInGround;
         }
 
         void implOnApplyTranslation(
             const OpenSim::PathPoint& pathPoint,
-            const Vec3& deltaTranslationInGround) final
+            const Vec3& translationInGround) final
         {
             const SimTK::Rotation parentToGroundRotation = pathPoint.getParentFrame().getRotationInGround(getState());
             const SimTK::InverseRotation& groundToParentRotation = parentToGroundRotation.invert();
-            const Vec3 translationInParent = ToVec3(groundToParentRotation * ToSimTKVec3(deltaTranslationInGround));
+            const Vec3 translationInParent = ToVec3(groundToParentRotation * ToSimTKVec3(translationInGround));
 
             ActionTranslatePathPoint(getUndoableModel(), pathPoint, translationInParent);
         }
@@ -300,7 +307,7 @@ namespace
         return SimTK::Rotation{-SimTK::Mat33{r}, true};
     }
 
-    // manipulator for `OpenSim::PhysicalOffsetFrame`
+    // an `ISelectionManipulator` that manipulates an `OpenSim::PhysicalOffsetFrame`
     class PhysicalOffsetFrameManipulator final : public SelectionManipulator<OpenSim::PhysicalOffsetFrame> {
     public:
         PhysicalOffsetFrameManipulator(
@@ -316,7 +323,7 @@ namespace
             return SupportedManipulationOpFlags::Translation | SupportedManipulationOpFlags::Rotation;
         }
 
-        Mat4 implGetCurrentModelMatrix(
+        Mat4 implGetCurrentTransformInGround(
             const OpenSim::PhysicalOffsetFrame& pof) const final
         {
             if (m_IsChildFrameOfJoint) {
@@ -335,32 +342,32 @@ namespace
 
         void implOnApplyTranslation(
             const OpenSim::PhysicalOffsetFrame& pof,
-            const Vec3& deltaTranslationInGround) final
+            const Vec3& translationInGround) final
         {
             SimTK::Rotation parentToGroundRotation = pof.getParentFrame().getRotationInGround(getState());
             if (m_IsChildFrameOfJoint) {
                 parentToGroundRotation = NegateRotation(parentToGroundRotation);
             }
             const SimTK::InverseRotation& groundToParentRotation = parentToGroundRotation.invert();
-            const SimTK::Vec3 deltaTranslationInParent = groundToParentRotation * ToSimTKVec3(deltaTranslationInGround);
+            const SimTK::Vec3 parentTranslation = groundToParentRotation * ToSimTKVec3(translationInGround);
             const SimTK::Vec3& eulersInPofFrame = pof.get_orientation();
 
             ActionTransformPof(
                 getUndoableModel(),
                 pof,
-                ToVec3(deltaTranslationInParent),
+                ToVec3(parentTranslation),
                 ToVec3(eulersInPofFrame)
             );
         }
 
         void implOnApplyRotation(
             const OpenSim::PhysicalOffsetFrame& pof,
-            const Eulers& deltaEulerRadiansInGround) final
+            const Eulers& eulersInLocalSpace) final
         {
             const OpenSim::Frame& parent = pof.getParentFrame();
             const SimTK::State& state = getState();
 
-            const Quat deltaRotationInGround = to_worldspace_rotation_quat(m_IsChildFrameOfJoint ? -deltaEulerRadiansInGround : deltaEulerRadiansInGround);
+            const Quat deltaRotationInGround = to_worldspace_rotation_quat(m_IsChildFrameOfJoint ? -eulersInLocalSpace : eulersInLocalSpace);
             const Quat oldRotationInGround = ToQuat(pof.getRotationInGround(state));
             const Quat parentRotationInGround = ToQuat(parent.getRotationInGround(state));
             const Quat newRotationInGround = normalize(deltaRotationInGround * oldRotationInGround);
@@ -372,13 +379,6 @@ namespace
                 Vec3{},  // no translation delta
                 extract_eulers_xyz(newRotationInParent)
             );
-        }
-
-        void implOnSave(const OpenSim::PhysicalOffsetFrame& pof) final
-        {
-            std::stringstream ss;
-            ss << "transformed " << pof.getName();
-            getUndoableModel().commit(std::move(ss).str());
         }
 
         bool m_IsChildFrameOfJoint = false;
@@ -399,7 +399,7 @@ namespace
             return SupportedManipulationOpFlags::Rotation | SupportedManipulationOpFlags::Translation;
         }
 
-        Mat4 implGetCurrentModelMatrix(
+        Mat4 implGetCurrentTransformInGround(
             const OpenSim::WrapObject& wrapObj) const final
         {
             const SimTK::Transform& wrapToFrame = wrapObj.getTransform();
@@ -411,11 +411,11 @@ namespace
 
         void implOnApplyTranslation(
             const OpenSim::WrapObject& wrapObj,
-            const Vec3& deltaTranslationInGround) final
+            const Vec3& translationInGround) final
         {
             const SimTK::Rotation frameToGroundRotation = wrapObj.getFrame().getTransformInGround(getState()).R();
             const SimTK::InverseRotation& groundToFrameRotation = frameToGroundRotation.invert();
-            const Vec3 translationInPofFrame = ToVec3(groundToFrameRotation * ToSimTKVec3(deltaTranslationInGround));
+            const Vec3 translationInPofFrame = ToVec3(groundToFrameRotation * ToSimTKVec3(translationInGround));
 
             ActionTransformWrapObject(
                 getUndoableModel(),
@@ -427,15 +427,16 @@ namespace
 
         void implOnApplyRotation(
             const OpenSim::WrapObject& wrapObj,
-            const Eulers& deltaEulerRadiansInGround) final
+            const Eulers& eulersInLocalSpace) final
         {
             const OpenSim::Frame& parent = wrapObj.getFrame();
             const SimTK::State& state = getState();
 
-            const Quat deltaRotationInGround = to_worldspace_rotation_quat(deltaEulerRadiansInGround);
+            const Quat deltaRotationInGround = to_worldspace_rotation_quat(eulersInLocalSpace);
             const Quat oldRotationInGround = ToQuat(parent.getTransformInGround(state).R() * wrapObj.getTransform().R());
-            const Quat parentRotationInGround = ToQuat(parent.getRotationInGround(state));
             const Quat newRotationInGround = normalize(deltaRotationInGround * oldRotationInGround);
+
+            const Quat parentRotationInGround = ToQuat(parent.getRotationInGround(state));
             const Quat newRotationInParent = inverse(parentRotationInGround) * newRotationInGround;
 
             ActionTransformWrapObject(
@@ -444,14 +445,6 @@ namespace
                 Vec3{},  // no translation
                 extract_eulers_xyz(newRotationInParent)
             );
-        }
-
-        void implOnSave(
-            const OpenSim::WrapObject& wrapObj) final
-        {
-            std::stringstream ss;
-            ss << "transformed " << wrapObj.getName();
-            getUndoableModel().commit(std::move(ss).str());
         }
     };
 
@@ -470,7 +463,7 @@ namespace
             return SupportedManipulationOpFlags::Rotation | SupportedManipulationOpFlags::Translation;
         }
 
-        Mat4 implGetCurrentModelMatrix(
+        Mat4 implGetCurrentTransformInGround(
             const OpenSim::ContactGeometry& contactGeom) const final
         {
             const SimTK::Transform wrapToFrame = contactGeom.getTransform();
@@ -505,8 +498,9 @@ namespace
 
             const Quat deltaRotationInGround = to_worldspace_rotation_quat(deltaEulerRadiansInGround);
             const Quat oldRotationInGround = ToQuat(parent.getTransformInGround(state).R() * contactGeom.getTransform().R());
-            const Quat parentRotationInGround = ToQuat(parent.getRotationInGround(state));
             const Quat newRotationInGround = normalize(deltaRotationInGround * oldRotationInGround);
+
+            const Quat parentRotationInGround = ToQuat(parent.getRotationInGround(state));
             const Quat newRotationInParent = inverse(parentRotationInGround) * newRotationInGround;
 
             ActionTransformContactGeometry(
@@ -516,152 +510,237 @@ namespace
                 extract_eulers_xyz(newRotationInParent)
             );
         }
+    };
 
-        void implOnSave(
-            const OpenSim::ContactGeometry& contactGeom) final
+    // an `ISelectionManipulator` that manipulates an `OpenSim::Joint` in the case where
+    // both sides of the joint are connected to `OpenSim::PhysicalOffsetFrame`s
+    class JointManipulator final : public SelectionManipulator<OpenSim::Joint> {
+    public:
+        static bool matches(const OpenSim::Component& component)
         {
-            std::stringstream ss;
-            ss << "transformed " << contactGeom.getName();
-            getUndoableModel().commit(std::move(ss).str());
+            const auto* joint = dynamic_cast<const OpenSim::Joint*>(&component);
+            if (not joint) {
+                return false;  // it isn't an `OpenSim::Joint`
+            }
+
+            if (not dynamic_cast<const OpenSim::PhysicalOffsetFrame*>(&joint->getParentFrame())) {
+                return false;  // the parent frame isn't an `OpenSim::PhysicalOffsetFrame`
+            }
+
+            if (not dynamic_cast<const OpenSim::PhysicalOffsetFrame*>(&joint->getChildFrame())) {
+                return false;  // the parent isn't an `OpenSim::PhysicalOffsetFrame`
+            }
+
+            return true;
+        }
+
+        JointManipulator(
+            std::shared_ptr<UndoableModelStatePair> model_,
+            const OpenSim::Joint& joint) :
+            SelectionManipulator{std::move(model_), joint}
+        {}
+
+    private:
+        SupportedManipulationOpFlags implGetSupportedManipulationOps() const final
+        {
+            return SupportedManipulationOpFlags::Translation | SupportedManipulationOpFlags::Rotation;
+        }
+
+        Mat4 implGetCurrentTransformInGround(const OpenSim::Joint& joint) const final
+        {
+            // present the "joint center" as equivalent to the parent frame
+            return ToMat4x4(joint.getParentFrame().getTransformInGround(getState()));
+        }
+
+        void implOnApplyTranslation(const OpenSim::Joint& joint, const Vec3& deltaTranslationInGround) final
+        {
+            applyTransform(joint, deltaTranslationInGround, Eulers{});
+        }
+
+        void implOnApplyRotation([[maybe_unused]] const OpenSim::Joint& joint, [[maybe_unused]] const Eulers& deltaEulersInLocalSpace) final
+        {
+            applyTransform(joint, Vec3{}, deltaEulersInLocalSpace);
+        }
+
+        void applyTransform(
+            const OpenSim::Joint& joint,
+            const Vec3& deltaTranslationInGround,
+            const Eulers& deltaEulersInLocalSpace)
+        {
+            // in order to move a joint center without every child also moving around, we need to:
+            //
+            // STEP 1) move the parent offset frame (as normal)
+            // STEP 2) figure out what transform the child offset frame would need to have in
+            //         order for its parent (confusing, eh) to not move
+            //
+            // this ended up being a headfuck, but I figured out that the easiest way to tackle this
+            // is to carefully track+name each frame definition and trust in god by using linear
+            // algebra to figure out the rest. So, given:
+            //
+            // - M_cpof1                    joint child offset frame to its parent transform (1: BEFORE)
+            // - M_j                        joint child-to-parent transform
+            // - M_ppof1                    joint parent offset frame to its parent transform (1: BEFORE)
+            // - M_ppof2                    joint parent offset frame to its parent transform (2: AFTER)
+            // - M_cpof2  **WE WANT THIS**  joint child offset frame to its parent transform (2: AFTER)
+            // - vcp                        an example quantity, expressed in the child's parent frame (e.g. a body)
+            // - vjp                        the same example quantity, but expressed in the joint's parent frame
+            //
+            // computing `vjp` from `vcp` involves going up the kinematic chain:
+            //
+            //     vjp = M_ppof1 * M_j * M_cpof1^-1 * vcp
+            //
+            // now, our goal is to apply STEP 1 (M_ppof1 --> M_ppof2) and calculate a new `M_cpof2` such that
+            // quantities expressed in a child body (e.g. `vcp`) do not move in the scene. I.e.:
+            //
+            //     vjp = M_ppof1 * M_j * M_cpof1^-1 * vcp = M_ppof2 * M_j * M_cpof2^-1 * vcp
+            //
+            // simplifying, and dropping the pretext of using the transforms to transform a particular point:
+            //
+            //     M_ppof1 * M_j * M_cpof1^-1 = M_ppof2 * M_j * M_cpof2^-1
+            //     M_ppof2^-1 * M_ppof1 * M_j * M_cpof1^-1 = M_j * M_cpof2^-1
+            //     M_j^-1 * M_ppof2^-1 * M_ppof1 * M_j * M_cpof1^-1 = M_cpof2^-1
+            //     M_cpof2^-1 = M_j^-1 * M_ppof2^-1 * M_ppof1 * M_j * M_cpof1^-1
+            //     M_cpof2 = (M_j^-1 * M_ppof2^-1 * M_ppof1 * M_j * M_cpof1^-1)^-1;
+            //
+            // the code below essentially collects all of this information up to figure out `M_cpof2` and stuff
+            // it into the child `OpenSim::PhysicalOffsetFrame`
+
+            const auto& parentPOF = dynamic_cast<const OpenSim::PhysicalOffsetFrame&>(joint.getParentFrame());
+            const auto& childPOF = dynamic_cast<const OpenSim::PhysicalOffsetFrame&>(joint.getChildFrame());
+
+            // get BEFORE transforms
+            const auto M_j = childPOF.findTransformBetween(getState(), parentPOF);
+            const auto M_ppof1 = parentPOF.findTransformBetween(getState(), parentPOF.getParentFrame());
+            const auto M_cpof1 = childPOF.findTransformBetween(getState(), childPOF.getParentFrame());
+
+            // STEP 1) move the parent offset frame (as normal)
+            {
+                const SimTK::Rotation R = ToSimTKRotation(deltaEulersInLocalSpace);
+                const auto& M_p = M_ppof1;
+                const auto M_o = parentPOF.getParentFrame().getTransformInGround(getState());
+
+                const SimTK::Rotation X_r = M_p.R() * R;
+                const SimTK::Vec3 X_p = M_p.p() + ToSimTKVec3(deltaTranslationInGround);
+                const SimTK::Transform X{X_r, X_p};
+                ActionTransformPof(
+                    getUndoableModel(),
+                    parentPOF,
+                    ToVec3(X.p() - M_ppof1.p()),
+                    ToVec3(X.R().convertRotationToBodyFixedXYZ())
+                );
+            }
+
+            // STEP 2) figure out what transform the child offset frame would need to have in
+            //         order for its parent (confusing, eh) to not move
+
+            // get AFTER transforms
+            const auto M_ppof2 = parentPOF.findTransformBetween(getState(), parentPOF.getParentFrame());
+
+            // caclulate `M_cpof2` (i.e. the desired new child transform)
+            const SimTK::Transform M_cpof2 = (M_j.invert() * M_ppof2.invert() * M_ppof1 * M_j * M_cpof1.invert()).invert();
+
+            // decompse `M_cpof2` into the child `OpenSim::PhysicalOffsetFrame`'s properties
+            ActionTransformPofV2(
+                getUndoableModel(),
+                childPOF,
+                ToVec3(M_cpof2.p()),
+                ToVec3(M_cpof2.R().convertRotationToBodyFixedXYZ())
+            );
         }
     };
+
+    // a compile-time `Typelist` containing all concrete implementations of
+    // `ISelectionManipulator`
+    using ManipulatorList = Typelist<
+        StationManipulator,
+        PathPointManipulator,
+        PhysicalOffsetFrameManipulator,
+        WrapObjectManipulator,
+        ContactGeometryManipulator,
+        JointManipulator
+    >;
 }
 
 // drawing/rendering code
 namespace
 {
+    // draws the gizmo overlay using the given `ISelectionManipulator`
     void DrawGizmoOverlayInner(
-        void* gizmoID,
+        ui::Gizmo& gizmo,
+        const Rect& screenRect,
         const PolarPerspectiveCamera& camera,
-        const Rect& viewportRect,
-        ImGuizmo::OPERATION operation,
-        ImGuizmo::MODE mode,
-        ISelectionManipulator& manipulator,
-        bool& wasUsingLastFrameStorage)
+        ISelectionManipulator& manipulator)
     {
         // figure out whether the gizmo should even be drawn
         {
             const SupportedManipulationOpFlags flags = manipulator.getSupportedManipulationOps();
-            if (operation == ImGuizmo::TRANSLATE && !(flags & SupportedManipulationOpFlags::Translation))
-            {
+            if (gizmo.operation() == ui::GizmoOperation::Translate && !(flags & SupportedManipulationOpFlags::Translation)) {
                 return;
             }
-            if (operation == ImGuizmo::ROTATE && !(flags & SupportedManipulationOpFlags::Rotation))
-            {
+            if (gizmo.operation() == ui::GizmoOperation::Rotate && !(flags & SupportedManipulationOpFlags::Rotation)) {
                 return;
             }
         }
-        // else: it's a supported operation and the gizmo should be drawn
 
-        // important: necessary for multi-viewport gizmos
-        // also important: don't use ui::get_id(), because it uses an ID stack and we might want to know if "isover" etc. is true outside of a window
-        ImGuizmo::SetID(static_cast<int>(std::hash<void*>{}(gizmoID)));
-        const ScopeGuard g{[]() { ImGuizmo::SetID(-1); }};
-
-        ImGuizmo::SetRect(
-            viewportRect.p1.x,
-            viewportRect.p1.y,
-            dimensions_of(viewportRect).x,
-            dimensions_of(viewportRect).y
+        // draw the manipulator
+        const Mat4 oldModelMatrix = manipulator.getCurrentTransformInGround();
+        Mat4 modelMatrix = oldModelMatrix;
+        const auto userEdit = gizmo.draw(
+            modelMatrix,
+            camera.view_matrix(),
+            camera.projection_matrix(aspect_ratio_of(screenRect)),
+            screenRect
         );
-        ImGuizmo::SetDrawlist(ui::get_panel_draw_list());
-        ImGuizmo::AllowAxisFlip(false);
+        if (userEdit) {
+            // propagate user edit to the model via the `ISelectionManipulator` interface
+            if (gizmo.operation() == ui::GizmoOperation::Translate) {
+                // `ISelectionManipulator` expects translations in ground
+                const Vec3 groundTranslation = transform_point(oldModelMatrix, userEdit->position);
+                manipulator.onApplyTranslation(groundTranslation);
+            }
+            else if (gizmo.operation() == ui::GizmoOperation::Rotate) {
+                manipulator.onApplyRotation(userEdit->rotation);
+            }
+        }
 
-        // use rotation from the parent, translation from station
-        Mat4 currentXformInGround = manipulator.getCurrentModelMatrix();
-        Mat4 deltaInGround;
-
-        ui::set_gizmo_style_to_osc_standard();
-        const bool gizmoWasManipulatedByUser = ImGuizmo::Manipulate(
-            value_ptr(camera.view_matrix()),
-            value_ptr(camera.projection_matrix(aspect_ratio_of(viewportRect))),
-            operation,
-            mode,
-            value_ptr(currentXformInGround),
-            value_ptr(deltaInGround),
-            nullptr,
-            nullptr,
-            nullptr
-        );
-
-        const bool isUsingThisFrame = ImGuizmo::IsUsing();
-        const bool wasUsingLastFrame = wasUsingLastFrameStorage;
-        wasUsingLastFrameStorage = isUsingThisFrame;  // update cached state
-
-        if (wasUsingLastFrame && !isUsingThisFrame)
-        {
-            // user finished interacting, save model
+        // once the user stops using the manipulator, save the changes
+        if (gizmo.was_using() and not gizmo.is_using()) {
             manipulator.onSave();
-        }
-
-        if (!gizmoWasManipulatedByUser)
-        {
-            return;  // user is not interacting, so no changes to apply
-        }
-        // else: apply in-place change to model
-
-        // decompose the overall transformation into component parts
-        Vec3 translationInGround{};
-        Vec3 rotationInGroundDegrees{};
-        Vec3 scaleInGround{};
-        ImGuizmo::DecomposeMatrixToComponents(
-            value_ptr(deltaInGround),
-            value_ptr(translationInGround),
-            value_ptr(rotationInGroundDegrees),
-            value_ptr(scaleInGround)
-        );
-        Eulers rotationInGround = Vec<3, Degrees>(rotationInGroundDegrees);
-
-        if (operation == ImGuizmo::TRANSLATE)
-        {
-            manipulator.onApplyTranslation(translationInGround);
-        }
-        else if (operation == ImGuizmo::ROTATE)
-        {
-            manipulator.onApplyRotation(rotationInGround);
         }
     }
 
-    void DrawGizmoOverlay(
-        void* gizmoID,
+    template<std::derived_from<ISelectionManipulator> ConcreteManipulator>
+    bool TryManipulateComponentWithManipulator(
+        ui::Gizmo& gizmo,
+        const Rect& screenRect,
         const PolarPerspectiveCamera& camera,
-        const Rect& viewportRect,
-        ImGuizmo::OPERATION operation,
-        ImGuizmo::MODE mode,
+        const std::shared_ptr<UndoableModelStatePair>& model,
+        const OpenSim::Component& selected)
+    {
+        if (not ConcreteManipulator::matches(selected)) {
+            return false;
+        }
+
+        const auto* downcasted = dynamic_cast<const typename ConcreteManipulator::AssociatedComponent*>(&selected);
+        if (not downcasted) {
+            return false;
+        }
+
+        ConcreteManipulator manipulator{model, *downcasted};
+        DrawGizmoOverlayInner(gizmo, screenRect, camera, manipulator);
+        return true;
+    }
+
+    template<typename... ConcreteManipulator>
+    void TryManipulateComponentWithMatchingManipulator(
+        ui::Gizmo& gizmo,
+        const Rect& screenRect,
+        const PolarPerspectiveCamera& camera,
         const std::shared_ptr<UndoableModelStatePair>& model,
         const OpenSim::Component& selected,
-        bool& wasUsingLastFrameStorage)
+        Typelist<ConcreteManipulator...>)
     {
-        // use downcasting to figure out which gizmo implementation to use
-        if (const auto* const maybeStation = dynamic_cast<const OpenSim::Station*>(&selected))
-        {
-            StationManipulator manipulator{model, *maybeStation};
-            DrawGizmoOverlayInner(gizmoID, camera, viewportRect, operation, mode, manipulator, wasUsingLastFrameStorage);
-        }
-        else if (const auto* const maybePathPoint = dynamic_cast<const OpenSim::PathPoint*>(&selected))
-        {
-            PathPointManipulator manipulator{model, *maybePathPoint};
-            DrawGizmoOverlayInner(gizmoID, camera, viewportRect, operation, mode, manipulator, wasUsingLastFrameStorage);
-        }
-        else if (const auto* const maybePof = dynamic_cast<const OpenSim::PhysicalOffsetFrame*>(&selected))
-        {
-            PhysicalOffsetFrameManipulator manipulator{model, *maybePof};
-            DrawGizmoOverlayInner(gizmoID, camera, viewportRect, operation, mode, manipulator, wasUsingLastFrameStorage);
-        }
-        else if (const auto* const maybeWrapObject = dynamic_cast<const OpenSim::WrapObject*>(&selected))
-        {
-            WrapObjectManipulator manipulator{model, *maybeWrapObject};
-            DrawGizmoOverlayInner(gizmoID, camera, viewportRect, operation, mode, manipulator, wasUsingLastFrameStorage);
-        }
-        else if (const auto* const maybeContactGeom = dynamic_cast<const OpenSim::ContactGeometry*>(&selected))
-        {
-            ContactGeometryManipulator manipulator{model, *maybeContactGeom};
-            DrawGizmoOverlayInner(gizmoID, camera, viewportRect, operation, mode, manipulator, wasUsingLastFrameStorage);
-        }
-        else
-        {
-            // (do nothing: we don't know how to manipulate the selection0
-        }
+        (TryManipulateComponentWithManipulator<ConcreteManipulator>(gizmo, screenRect, camera, model, selected) || ...);
     }
 }
 
@@ -674,45 +753,21 @@ osc::ModelSelectionGizmo& osc::ModelSelectionGizmo::operator=(const ModelSelecti
 osc::ModelSelectionGizmo& osc::ModelSelectionGizmo::operator=(ModelSelectionGizmo&&) noexcept = default;
 osc::ModelSelectionGizmo::~ModelSelectionGizmo() noexcept = default;
 
-bool osc::ModelSelectionGizmo::isUsing() const
-{
-    ImGuizmo::SetID(static_cast<int>(std::hash<const ModelSelectionGizmo*>{}(this)));
-    const bool rv = ImGuizmo::IsUsing();
-    ImGuizmo::SetID(-1);
-    return rv;
-}
-
-bool osc::ModelSelectionGizmo::isOver() const
-{
-    ImGuizmo::SetID(static_cast<int>(std::hash<const ModelSelectionGizmo*>{}(this)));
-    const bool rv = ImGuizmo::IsOver();
-    ImGuizmo::SetID(-1);
-    return rv;
-}
-
-bool osc::ModelSelectionGizmo::handleKeyboardInputs()
-{
-    return ui::update_gizmo_state_from_keyboard(m_GizmoOperation, m_GizmoMode);
-}
-
 void osc::ModelSelectionGizmo::onDraw(
     const Rect& screenRect,
     const PolarPerspectiveCamera& camera)
 {
     const OpenSim::Component* selected = m_Model->getSelected();
-    if (!selected)
-    {
+    if (not selected) {
         return;
     }
 
-    DrawGizmoOverlay(
-        this,
-        camera,
+    TryManipulateComponentWithMatchingManipulator(
+        m_Gizmo,
         screenRect,
-        m_GizmoOperation,
-        m_GizmoMode,
+        camera,
         m_Model,
         *selected,
-        m_WasUsingGizmoLastFrame
+        ManipulatorList{}
     );
 }