+++ src/oscar/UI.h
@@ -8,7 +8,6 @@
 #include <oscar/UI/Icon.h>
 #include <oscar/UI/IconCache.h>
 #include <oscar/UI/ImGuiHelpers.h>
-#include <oscar/UI/ImGuizmoHelpers.h>
 #include <oscar/UI/MouseCapturingCamera.h>
 #include <oscar/UI/oscimgui.h>
 #include <oscar/UI/oscimgui_config.h>